Arreglillos en la seguridad

+++ MeowCafe02_SergioGonzalezCortes/src/main/java/com/salesianostriana/dam/meowcafe02sergiogonzalezcortes/model/Combo.java
@@ -31,7 +31,7 @@ public class Combo {
 	
 	private boolean esOferta;
 	
-	private float precioCombo;
+	private double precioCombo;
 	
 	@ToString.Exclude
 	@EqualsAndHashCode.Exclude
